+++ changelog.md
@@ -1,6 +1,7 @@
 ## [0.0.2] - 2022-22-02 - R.S.
 ### Ajout
 -Mise en place de la réponse au message.
+-Mise en place de la création et connexion à des lobbys.
 
 ## [0.0.1] - 2022-21-02 - R.S.
 ### Ajout